Refactor Cell STS to intercept both inbound and outbound calls within the cell.
