+++ system/control-plane/cell/components/cell-sts/src/main/java/org/wso2/vick/auth/cell/sts/VickCellSTSServer.java
@@ -2,9 +2,13 @@ package org.wso2.vick.auth.cell.sts;
 
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
+import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.vick.auth.cell.sts.service.VickCellAuthorizationService;
+import org.wso2.vick.auth.cell.sts.context.store.UserContextStore;
+import org.wso2.vick.auth.cell.sts.context.store.UserContextStoreImpl;
+import org.wso2.vick.auth.cell.sts.service.VickCellInboundInterceptorService;
+import org.wso2.vick.auth.cell.sts.service.VickCellOutboundInterceptorService;
 import org.wso2.vick.auth.cell.sts.service.VickCellSTSException;
 
 import java.io.IOException;
@@ -15,13 +19,25 @@ import java.io.IOException;
 public class VickCellSTSServer {
 
     private static final Logger log = LoggerFactory.getLogger(VickCellSTSServer.class);
-    private final int port;
-    private final Server server;
+    private final int inboundListeningPort;
+    private final Server inboundListener;
 
-    private VickCellSTSServer(int port) throws VickCellSTSException {
+    private final int outboundListeningPort;
+    private final Server outboundListener;
 
-        this.port = port;
-        server = ServerBuilder.forPort(port).addService(new VickCellAuthorizationService()).build();
+    private VickCellSTSServer(int inboundListeningPort, int outboundListeningPort) throws VickCellSTSException {
+
+        UserContextStore contextStore = new UserContextStoreImpl();
+
+        this.inboundListeningPort = inboundListeningPort;
+        inboundListener = ServerBuilder.forPort(inboundListeningPort)
+                .addService(new VickCellInboundInterceptorService(contextStore))
+                .build();
+
+        this.outboundListeningPort = outboundListeningPort;
+        outboundListener = ServerBuilder.forPort(outboundListeningPort)
+                .addService(new VickCellOutboundInterceptorService(contextStore))
+                .build();
     }
 
     /**
@@ -29,8 +45,10 @@ public class VickCellSTSServer {
      */
     private void start() throws IOException {
 
-        server.start();
-        log.info("Vick Cell STS GRPC Server started, listening on " + port);
+        inboundListener.start();
+        outboundListener.start();
+        log.info("Vick Cell STS GRPC Server started, listening for inbound traffic on " + inboundListeningPort);
+        log.info("Vick Cell STS GRPC Server started, listening for outbound traffic on " + outboundListeningPort);
         Runtime.getRuntime().addShutdownHook(new Thread(() -> {
             // Use stderr here since the logger may has been reset by its JVM shutdown hook.
             System.err.println("Shutting down Vick Cell STS since JVM is shutting down.");
@@ -44,8 +62,12 @@ public class VickCellSTSServer {
      */
     private void stop() {
 
-        if (server != null) {
-            server.shutdown();
+        if (inboundListener != null) {
+            inboundListener.shutdown();
+        }
+
+        if (outboundListener != null) {
+            outboundListener.shutdown();
         }
     }
 
@@ -54,16 +76,20 @@ public class VickCellSTSServer {
      */
     private void blockUntilShutdown() throws InterruptedException {
 
-        if (server != null) {
-            server.awaitTermination();
+        if (inboundListener != null) {
+            inboundListener.awaitTermination();
+        }
+
+        if (outboundListener != null) {
+            outboundListener.awaitTermination();
         }
     }
 
     public static void main(String[] args) {
 
-        VickCellSTSServer server = null;
+        VickCellSTSServer server;
         try {
-            server = new VickCellSTSServer(8080);
+            server = new VickCellSTSServer(8080, 8081);
             server.start();
             server.blockUntilShutdown();
         } catch (Exception e) {
