+++ system/control-plane/cell/components/cell-sts/src/main/resources/sts.json
@@ -1,5 +1,5 @@
 {
   "endpoint": "https://localhost:9443/api/identity/vick-auth/v1.0/sts/token",
-  "username": "####",
-  "password": "####"
+  "username": "admin",
+  "password": "admin"
 }
